python-clip-service/requirements.txt
flask
flask-cors
torch
torchvision
transformers
pillow
numpy
requests
python-dotenv
gunicorn
